Rename DailyAvailability to Availability and update related configurations

+++ travel-agency-query-side/src/main/java/com/rzodeczko/application/port/in/UpsertHotelCapacityUseCase.java
@@ -1,5 +1,5 @@
 package com.rzodeczko.application.port.in;
 
 public interface UpsertHotelCapacityUseCase {
-    void upsert(long hotelId, int capacity);
+    void upsert(long hotelId, long capacity);
 }
